:truck: move `StudentForEnrolment` to `src/certification/enrolment/`
