+++ api/src/certification/enrolment/domain/read-models/StudentForEnrolment.js
