+++ api/tests/certification/enrolment/unit/infrastructure/serializers/student-certification-serializer_test.js
@@ -1,5 +1,5 @@
+import { StudentForEnrolment } from '../../../../../../src/certification/enrolment/domain/read-models/StudentForEnrolment.js';
 import * as serializer from '../../../../../../src/certification/enrolment/infrastructure/serializers/student-certification-serializer.js';
-import { StudentForEnrolment } from '../../../../../../src/shared/domain/read-models/StudentForEnrolment.js';
 import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | student-certification-serializer', function () {
